agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
copilotkit>=0.1.30
ag-ui-langgraph>=0.0.24
tavily-python>=0.5.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
earthengine-api>=1.2.0
httpx>=0.28.0
pydantic>=2.0.0
requests>=2.31.0
